refactor: Provider 삭제 및 로직 이동
